requirements.txt
pandas

langchain
langchain_community


langchain_groq

chromadb

python-dotenv

streamlit
